+++ Minecraft Clone/src/RenderEngine/EntityRenderer/Models/EntityModel.h
@@ -69,7 +69,7 @@ public:
             }
 
             for (size_t i = 0; i < SubModel.indices_.size(); i++) {
-                model.indices_.push_back(SubModel.indices_[i] + CurrentIndex);
+                model.indices_.push_back(static_cast<uint32_t>(SubModel.indices_[i] + CurrentIndex));
             }
 
             model.vertices_count_ += SubModel.vertices_count_;
